@@ -3,7 +3,7 @@ import 'dart:math';
 class CalculatorBrain {
   final int height;
   final int weight;
-  double _bmi;
+  late double _bmi;
 
   CalculatorBrain(this.height, this.weight);
 
@@ -28,7 +28,7 @@ class CalculatorBrain {
     } else if (_bmi > 18.5) {
       return 'You have a normal body weight. Good job!';
     } else {
-      return 'You have a lower than normal bodu wight. You can eat a bit more.';
+      return 'You have a lower than normal body weight. You can eat a bit more.';
     }
   }
 }
