@@ -2,16 +2,16 @@ import 'package:flutter/material.dart';
 
 class RoundIconButton extends StatelessWidget {
   final IconData icon;
-  final Function onPressed;
+  final VoidCallback? onPressed;
 
-  RoundIconButton({this.icon, this.onPressed});
+  const RoundIconButton({super.key, required this.icon, this.onPressed});
 
   @override
   Widget build(BuildContext context) {
     return RawMaterialButton(
       elevation: 0,
-      child: Icon(icon),
       onPressed: onPressed,
+      child: Icon(icon),
       constraints: BoxConstraints.tightFor(
         width: 50,
         height: 50,
