@@ -8,9 +8,13 @@ class MyApp extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
     return MaterialApp(
+      debugShowCheckedModeBanner: false,
       theme: ThemeData.dark().copyWith(
-        primaryColor: Color(0xFF0A0E21),
-        scaffoldBackgroundColor: Color(0xFF0A0E21),
+        colorScheme: ThemeData.dark().colorScheme.copyWith(
+              primary: const Color(0xFF0A0E21),
+              surface: const Color(0xFF0A0E21),
+            ),
+        scaffoldBackgroundColor: const Color(0xFF0A0E21),
       ),
       home: InputPage(),
     );
